Replace TMessageDialog with TMessageList (unfinished)
The new TMessageList is cleaner in the sense that the list of messages
is not part of a dialog, but is distinct from the user interface. It
also contains a wealth of methods that make it possible to the caller to
avoid many "if"s. However, I am still unsatisfied with its design and I
am temporarily (?) discontinuing this refactoring because it is taking
longer than I thought. It is almost finished, however.